- Add getTokenPrice hook. - Separate minimalpositioncard and position card into different files, with separate styles and types files - Add customSwap parameters for userDerivedSwapInfo - TODO: Fix rpc problems when fetching token price too fast
