+++ src/components/PositionCard/PositionCard.types.ts
@@ -1,0 +1,7 @@
+import { Pair } from '@trisolaris/sdk'
+
+export interface PositionCardProps {
+  pair: Pair
+  showUnwrapped?: boolean
+  border?: string
+}
