+++ src/components/PositionCard/V1.tsx
@@ -6,7 +6,7 @@ import { Text } from 'rebass'
 import { AutoColumn } from '../Column'
 import { ButtonSecondary } from '../Button'
 import { RowBetween, RowFixed } from '../Row'
-import { FixedHeightRow, HoverCard } from './index'
+import { FixedHeightRow, HoverCard } from './PositionCard.styles'
 import DoubleCurrencyLogo from '../DoubleLogo'
 import { useActiveWeb3React } from '../../hooks'
 import { ThemeContext } from 'styled-components'
