+++ src/components/PositionCard/index.tsx
@@ -1,167 +1,32 @@
-import { JSBI, Pair, Percent } from '@trisolaris/sdk'
-import { darken } from 'polished'
 import React, { useState } from 'react'
+import { JSBI, Percent, Fraction } from '@trisolaris/sdk'
 import { ChevronDown, ChevronUp } from 'react-feather'
 import { Link } from 'react-router-dom'
 import { Text } from 'rebass'
-import styled from 'styled-components'
-import { useTotalSupply } from '../../data/TotalSupply'
+import { useTranslation } from 'react-i18next'
 
-import { useActiveWeb3React } from '../../hooks'
-import { useTokenBalance } from '../../state/wallet/hooks'
-import { TYPE } from '../../theme'
-import { currencyId } from '../../utils/currencyId'
-import { unwrappedToken } from '../../utils/wrappedCurrency'
-import { ButtonPrimary, ButtonEmpty } from '../Button'
-import { TokenPairBackgroundColor } from '../earn/PoolCardTri.styles'
-
-import { useColorWithDefault } from '../../hooks/useColor'
-
-import Card, { GreyCard, LightCard } from '../Card'
+import { ButtonPrimary } from '../Button'
 import { AutoColumn } from '../Column'
 import CurrencyLogo from '../CurrencyLogo'
 import DoubleCurrencyLogo from '../DoubleLogo'
 import { RowBetween, RowFixed } from '../Row'
 import { Dots } from '../swap/styleds'
-import { useTranslation } from 'react-i18next'
+
+import { useTotalSupply } from '../../data/TotalSupply'
+import { useActiveWeb3React } from '../../hooks'
+import { useTokenBalance } from '../../state/wallet/hooks'
+import { useColorWithDefault } from '../../hooks/useColor'
 import { getPairRenderOrder } from '../../utils/pools'
+import useGetTokenPrice from '../../hooks/useGetTokenPrice'
+
+import { currencyId } from '../../utils/currencyId'
 
 import { BIG_INT_ZERO } from '../../constants'
+import { PositionCardProps } from './PositionCard.types'
 
-export const FixedHeightRow = styled(RowBetween)`
-  height: 24px;
-`
+import { TokenPairBackgroundColor } from '../earn/PoolCardTri.styles'
 
-export const HoverCard = styled(Card)`
-  border: 1px solid transparent;
-  :hover {
-    border: 1px solid ${({ theme }) => darken(0.06, theme.bg2)};
-  }
-`
-const StyledPositionCard = styled(LightCard)<{ bgColor: any }>`
-  background-color: unset;
-  border: none;
-  position: relative;
-  overflow: hidden;
-`
-
-interface PositionCardProps {
-  pair: Pair
-  showUnwrapped?: boolean
-  border?: string
-}
-
-const ManageButton = styled(ButtonEmpty)`
-  color: ${({ theme }) => theme.white};
-`
-
-export function MinimalPositionCard({ pair, showUnwrapped = false, border }: PositionCardProps) {
-  const { account } = useActiveWeb3React()
-
-  const currency0 = showUnwrapped ? pair.token0 : unwrappedToken(pair.token0)
-  const currency1 = showUnwrapped ? pair.token1 : unwrappedToken(pair.token1)
-
-  const { t } = useTranslation()
-  const [showMore, setShowMore] = useState(false)
-
-  const userPoolBalance = useTokenBalance(account ?? undefined, pair.liquidityToken)
-  const totalPoolTokens = useTotalSupply(pair.liquidityToken)
-
-  const poolTokenPercentage =
-    !!userPoolBalance && !!totalPoolTokens && JSBI.greaterThanOrEqual(totalPoolTokens.raw, userPoolBalance.raw)
-      ? new Percent(userPoolBalance.raw, totalPoolTokens.raw)
-      : undefined
-
-  const [token0Deposited, token1Deposited] =
-    !!pair &&
-    !!totalPoolTokens &&
-    !!userPoolBalance &&
-    // this condition is a short-circuit in the case where useTokenBalance updates sooner than useTotalSupply
-    JSBI.greaterThanOrEqual(totalPoolTokens.raw, userPoolBalance.raw)
-      ? [
-          pair.getLiquidityValue(pair.token0, totalPoolTokens, userPoolBalance, false),
-          pair.getLiquidityValue(pair.token1, totalPoolTokens, userPoolBalance, false)
-        ]
-      : [undefined, undefined]
-
-  return (
-    <>
-      {userPoolBalance && JSBI.greaterThan(userPoolBalance.raw, BIG_INT_ZERO) ? (
-        <GreyCard border={border}>
-          <AutoColumn gap="12px">
-            <FixedHeightRow>
-              <RowFixed>
-                <Text fontWeight={500} fontSize={16}>
-                  {t('positionCard.yourPosition')}
-                </Text>
-              </RowFixed>
-            </FixedHeightRow>
-            <FixedHeightRow onClick={() => setShowMore(!showMore)}>
-              <RowFixed>
-                <DoubleCurrencyLogo currency0={currency0} currency1={currency1} margin={true} size={20} />
-                <Text fontWeight={500} fontSize={20}>
-                  {currency0.symbol}/{currency1.symbol}
-                </Text>
-              </RowFixed>
-              <RowFixed>
-                <Text fontWeight={500} fontSize={20}>
-                  {userPoolBalance ? userPoolBalance.toSignificant(4) : '-'}
-                </Text>
-              </RowFixed>
-            </FixedHeightRow>
-            <AutoColumn gap="4px">
-              <FixedHeightRow>
-                <Text fontSize={16} fontWeight={500}>
-                  {t('positionCard.poolShare')}
-                </Text>
-                <Text fontSize={16} fontWeight={500}>
-                  {poolTokenPercentage ? poolTokenPercentage.toFixed(6) + '%' : '-'}
-                </Text>
-              </FixedHeightRow>
-              <FixedHeightRow>
-                <Text fontSize={16} fontWeight={500}>
-                  {currency0.symbol}:
-                </Text>
-                {token0Deposited ? (
-                  <RowFixed>
-                    <Text fontSize={16} fontWeight={500} marginLeft={'6px'}>
-                      {token0Deposited?.toSignificant(6)}
-                    </Text>
-                  </RowFixed>
-                ) : (
-                  '-'
-                )}
-              </FixedHeightRow>
-              <FixedHeightRow>
-                <Text fontSize={16} fontWeight={500}>
-                  {currency1.symbol}:
-                </Text>
-                {token1Deposited ? (
-                  <RowFixed>
-                    <Text fontSize={16} fontWeight={500} marginLeft={'6px'}>
-                      {token1Deposited?.toSignificant(6)}
-                    </Text>
-                  </RowFixed>
-                ) : (
-                  '-'
-                )}
-              </FixedHeightRow>
-            </AutoColumn>
-          </AutoColumn>
-        </GreyCard>
-      ) : (
-        <LightCard>
-          <TYPE.subHeader style={{ textAlign: 'center' }}>
-            <span role="img" aria-label="wizard-icon">
-              ⭐️
-            </span>{' '}
-            {t('positionCard.byAddingLiquidityInfo')}
-          </TYPE.subHeader>
-        </LightCard>
-      )}
-    </>
-  )
-}
+import { ManageButton, FixedHeightRow, StyledPositionCard } from './PositionCard.styles'
 
 export default function FullPositionCard({ pair, border }: PositionCardProps) {
   const { account } = useActiveWeb3React()
@@ -192,6 +57,17 @@ export default function FullPositionCard({ pair, border }: PositionCardProps) {
 
   const backgroundColor1 = useColorWithDefault('#2172E5', token0)
   const backgroundColor2 = useColorWithDefault('#2172E5', token1)
+
+  // const token0Price = useGetTokenPrice(token0.address)
+  // const token1Price = useGetTokenPrice(token1.address)
+
+  // const token0Deposits =
+  //   token0Deposited?.multiply(token0Price?.adjusted ?? JSBI.BigInt(1)) ?? new Fraction(BIG_INT_ZERO)
+  // const token1Deposits =
+  //   token1Deposited?.multiply(token1Price?.adjusted ?? JSBI.BigInt(1)) ?? new Fraction(BIG_INT_ZERO)
+
+  // const totalDeposits = token0Deposits.add(token1Deposits)
+
 
   return (
     <StyledPositionCard border={border} bgColor={backgroundColor1}>
@@ -242,6 +118,15 @@ export default function FullPositionCard({ pair, border }: PositionCardProps) {
                 {userPoolBalance ? userPoolBalance.toSignificant(4) : '-'}
               </Text>
             </FixedHeightRow>
+            <FixedHeightRow>
+              <Text fontSize={16} fontWeight={500}>
+                Your deposits
+              </Text>
+              <Text fontSize={16} fontWeight={500}>
+                {/* {totalDeposits.greaterThan(BIG_INT_ZERO) ? `$${totalDeposits.toSignificant(6)}` : '-'} */}
+              </Text>
+            </FixedHeightRow>
+
             <FixedHeightRow>
               <RowFixed>
                 <Text fontSize={16} fontWeight={500}>
