+++ src/pages/RemoveLiquidity/index.tsx
@@ -19,8 +19,8 @@ import AppBody from '../AppBody'
 import { ButtonPrimary, ButtonLight, ButtonError, ButtonConfirmed } from '../../components/Button'
 import { AutoColumn, ColumnCenter } from '../../components/Column'
 import { AddRemoveTabs } from '../../components/NavigationTabs'
-import { MinimalPositionCard } from '../../components/PositionCard'
-import { RowBetween, RowFixed } from '../../components/Row'
+import MinimalPositionCard from '../../components/PositionCard'
+import Row, { RowBetween, RowFixed } from '../../components/Row'
 
 import useTransactionDeadline from '../../hooks/useTransactionDeadline'
 import useDebouncedChangeHandler from '../../utils/useDebouncedChangeHandler'
