+++ src/state/swap/hooks.ts
@@ -115,8 +115,10 @@ function involvesAddress(trade: Trade, checksummedAddress: string): boolean {
   )
 }
 
-// from the current swap inputs, compute the best trade and return it.
-export function useDerivedSwapInfo(): {
+// By default, from the current swap inputs, compute the best trade and return it. Optionally, receives a custom SwapState
+export function useDerivedSwapInfo(
+  customSwap?: SwapState
+): {
   currencies: { [field in Field]?: Currency }
   currencyBalances: { [field in Field]?: CurrencyAmount }
   parsedAmount: CurrencyAmount | undefined
@@ -129,13 +131,14 @@ export function useDerivedSwapInfo(): {
 
   const toggledVersion = useToggledVersion()
 
+  const swapData = customSwap ?? useSwapState()
   const {
     independentField,
     typedValue,
     [Field.INPUT]: { currencyId: inputCurrencyId },
     [Field.OUTPUT]: { currencyId: outputCurrencyId },
     recipient
-  } = useSwapState()
+  } = swapData
 
   const inputCurrency = useCurrency(inputCurrencyId)
   const outputCurrency = useCurrency(outputCurrencyId)
